requirements.txt
torch>=2.0.0
torchvision>=0.15.0
medmnist>=2.2.0
numpy>=1.24.0
pandas>=2.0.0
matplotlib>=3.7.0
seaborn>=0.12.0
scikit-learn>=1.2.0
tqdm>=4.65.0
pyyaml>=6.0
httpx
ollama
Pillow
fastapi
uvicorn[standard]
sqlalchemy>=2.0
pgvector
psycopg2-binary
python-dotenv
python-multipart
transformers>=4.45.0
accelerate>=0.26.0
ollama
httpx